requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
streamlit>=1.40.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.55.0,<2.0.0
google-generativeai>=0.8.0,<1.0.0
pypdf>=5.0.0,<6.0.0
python-multipart>=0.0.9,<1.0.0
faiss-cpu>=1.8.0,<2.0.0
numpy>=1.26.0,<3.0.0
requests>=2.31.0,<3.0.0
pymupdf>=1.24.0,<2.0.0
Pillow>=10.0.0,<12.0.0
imageio>=2.33.0,<3.0.0
imageio-ffmpeg>=0.5.0,<1.0.0
